Assignment-1/requirements.txt
fastapi
uvicorn
langchain
openai
python-dotenv
pydantic
faiss-cpu
sentence-transformers
transformers
requests
python-multipart